% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtextgrid.R
\name{read_textgrid}
\alias{read_textgrid}
\alias{read_textgrid_lines}
\title{Read a textgrid file into a tibble}
\usage{
read_textgrid(path)

read_textgrid_lines(lines, file = NA_character_)
}
\arguments{
\item{path}{a path to a dataframe}

\item{lines}{alternatively, the lines of a textgrid file}
}
\value{
a tibble with one row per textgrid annotation
}
\description{
Read a textgrid file into a tibble
}
