% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtextgrid-package.R
\docType{package}
\name{readtextgrid-package}
\alias{readtextgrid}
\alias{readtextgrid-package}
\title{readtextgrid: Read in a 'Praat' 'TextGrid' File}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

'Praat' \url{https://www.fon.hum.uva.nl/praat/} is a widely used tool for manipulating, annotating and analyzing speech and acoustic data. It stores annotation data in a format called a 'TextGrid'. This package provides a way to read these files into R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjmahr/readtextgrid}
  \item \url{https://www.tjmahr.com/readtextgrid/}
  \item Report bugs at \url{https://github.com/tjmahr/readtextgrid/issues}
}

}
\author{
\strong{Maintainer}: Tristan Mahr \email{tristan.mahr@wisc.edu} (\href{https://orcid.org/0000-0002-8890-5116}{ORCID})

Authors:
\itemize{
  \item Josef Fruehwald
}

Other contributors:
\itemize{
  \item Dan Villarreal [contributor]
  \item Jonathan Washington [contributor]
}

}
\keyword{internal}
