% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtextgrid.R
\name{example_textgrid}
\alias{example_textgrid}
\title{Locate the path of the example textgrid file}
\usage{
example_textgrid()
}
\value{
Path of \code{"Mary_John_bell.TextGrid"} bundled with the \code{readtextgrid}
package.
}
\description{
Locate the path of the example textgrid file
}
\details{
This function is a wrapper over \code{\link[=system.file]{system.file()}}  to locate the
path to \code{"Mary_John_bell.TextGrid"}. This file is the default textgrid that
is created by Praat.
}
